Code refactoring.

+++ Poker/Interfaces/IPlayer.cs
@@ -28,7 +28,7 @@
 
         Label Status { get; set; }
 
-        void FixCall(ref int globalRaise, ref double globalCall, int options, double globalRounds, ref Button callButton);
+        void FixCall(ref int globalRaise, ref int globalCall, int options, int globalRounds, ref Button callButton);
 
         void Rules();
     }
